textsummarizer.py


import gensim
from gensim.summarization import summarize
import torch
from transformers import pipeline, T5Tokenizer, T5ForConditionalGeneration, T5Config
from summarizer import Summarizer, TransformerSummarizer
