requirements.txt
bert-extractive-summarizer==0.10.1
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8        
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cymem==2.0.6
Cython==0.29.14
filelock==3.6.0
flake8==4.0.1
gensim==3.8.3
huggingface-hub==0.5.0    
idna==3.3
importlib-metadata==4.2.0 
Jinja2==3.1.1
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
mccabe==0.6.1
murmurhash==1.0.6
numpy==1.21.5
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
pyparsing==3.0.7
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
tqdm==4.64.0
transformers==4.17.0
typer==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.9
wasabi==0.9.1
zipp==3.8.0